Replace string.split with text.split
Not only is modifying builtins seen as bad (though I will do it
when it makes sense in lennie), but this implementation of split
does weird things like capturing the line number that the splits
ocur on! This is useful for the Template library but not generic
enough to put on the builtin string table.

Instead I've introduced the text table locally. I'm imaging that
the difference between string and text for me is that in text,
newlines are important where as in strings they are just another
character. This makes the bahavior of capuring line numbers as well a
little more intuative.