Extract and Expose Server:handle function and use it in :run
When testing a server I don't want to have to spin up a separate thread
to run the server and then communicate with it over a socket. This is
something that I have experienced in rails and other web frameworks. Not
only is it annoying but it can cause issues with managing database
access (not that we have that yet). Instead I have exposed the :handle
function which will process a request and return a response. Then in
:run we can construct a request from a socket connection, pass it to
:handle and render the response.

In the tests we can do a number of things then:

1. Access the raw response. I'm also preserving the body object returned
   so that tests can query the content of the view object. No more
   brittle UI tests!

2. If you need brittle UI tests (and sometimes we do!) then you can
   render the view object in the test and make assertions against it.
   You could even load the result of the view object into a headless
   browser to test! One issue here is that without a sever running none
   of the linked resources would be requested... perhaps there is value
   in having a server running. Or perhaps these type of tests could have
   a separate test harness?