Add capturing matches to routing
This allows users to follow a common pattern for identifying parts of a
route to match which act as parameters to handling the request. A
example might be:

route '/person/:name'

Here ':name' represents a part of the url which we will required to
serve the request.

I'm opting not to introduce a new string matching language here and just
to use patterns from luas standard library. Matching libraries could
implement this feature on their own! Though I will probably implement
such a library.

Instead we will simply pass through captures from the patterns as
parameters to the route handler e.g:

```
router:get('^/person/([^/]+)$', function (state, request, name)
  ...
end)
```

This isn't as clear as the above but is good enough for now.