Fix lazy loading values in templates.
Previously, we were making "holes" in the template with '%s' which would
later be filled by functions. Unfortunately other code and users could
simply have '%s' or other characters embedded in the template. This
would break the use of string formatting.

Instead I've moved for simply adding the lazy loaded values as functions
in the table of lines and when it comes to combining the lines into the
final template I evaluate those functions.

No more relying on the template formatting to know where to put lazy
values!